Yawn. This took me 3 hours just to compile on Windows. But it's done. Until next week, I guess.

+++ libinca/src/inca/integration/win32/win32-missing.cpp
@@ -23,7 +23,7 @@ float roundf(float x) {
 }
 
 // ...or for doubles...
-double roundl(double x) {
+double round(double x) {
     return (double)long(x);
 }
